Move the changes to feehandler to proper module
