+++ core-contracts/FeeHandler/src/main/java/network/balanced/score/core/feehandler/FeeHandlerImpl.java
@@ -41,6 +41,9 @@ public class FeeHandlerImpl implements FeeHandler {
     private final String LAST_TXHASH = "last_txhash";
     private final String ROUTES = "routes";
     private final String GOVERNANCE = "governance";
+    private final String LOANS = "loans";
+    private final String DEX = "dex";
+    private final String STABILITY = "stability";
     private final String ENABLED = "enabled";
     private final String ALLOWED_ADDRESS = "allowed_address";
     private final String NEXT_ALLOWED_ADDRESS_INDEX = "_next_allowed_addresses_index";
@@ -52,11 +55,17 @@ public class FeeHandlerImpl implements FeeHandler {
     private final VarDB<byte[]> lastTxhash = Context.newVarDB(LAST_TXHASH, byte[].class);
     private final BranchDB<Address, DictDB<Address, String>> routes = Context.newBranchDB(ROUTES, String.class);
     private final VarDB<Address> governance = Context.newVarDB(GOVERNANCE, Address.class);
+    private final VarDB<Address> loans = Context.newVarDB(LOANS, Address.class);
+    private final VarDB<Address> dex = Context.newVarDB(DEX, Address.class);
+    private final VarDB<Address> stabilityFund = Context.newVarDB(STABILITY, Address.class);
     private final VarDB<Boolean> enabled = Context.newVarDB(ENABLED, Boolean.class);
     private final ArrayDB<Address> allowedAddress = Context.newArrayDB(ALLOWED_ADDRESS, Address.class);
     private final VarDB<Integer> nextAllowedAddressesIndex = Context.newVarDB(NEXT_ALLOWED_ADDRESS_INDEX,
             Integer.class);
     private final VarDB<Address> admin = Context.newVarDB(ADMIN_ADDRESS, Address.class);
+    private final DictDB<Address, BigInteger> swapFeesAccruedDB = Context.newDictDB(DIVIDEND_TOKENS, BigInteger.class);
+    private BigInteger loanFeesAccrued;
+    private BigInteger stabilityFundFeesAccrued;
 
     public FeeHandlerImpl(Address _governance) {
         if (governance.get() == null) {
@@ -80,6 +89,56 @@ public class FeeHandlerImpl implements FeeHandler {
     @External(readonly = true)
     public Address getGovernance() {
         return governance.get();
+    }
+
+    // acceptedDividendsTokens DB is used here to add the token types of fees generated by the DEX to swapFees DB
+    @External
+    public void setSwapFeesAccruedDB() {
+        for (int i = 0; i < acceptedDividendsTokens.size(); i++) {
+            Address thisToken = acceptedDividendsTokens.get(i);
+            swapFeesAccruedDB.set(thisToken, swapFeesAccruedDB.get(thisToken));
+        }
+    }
+
+    @External(readonly = true)
+    public BigInteger getSwapFeesAccruedByToken(Address token) {
+        return swapFeesAccruedDB.get(token);
+    }
+
+    @External
+    public void setLoans(Address _address) {
+        onlyOwner();
+        isContract(_address);
+        loans.set(_address);
+    }
+
+    @External(readonly = true)
+    public Address getLoans() {
+        return loans.get();
+    }
+
+    @External
+    public void setDex(Address _address) {
+        onlyOwner();
+        isContract(_address);
+        dex.set(_address);
+    }
+
+    @External(readonly = true)
+    public Address getDex() {
+        return dex.get();
+    }
+
+    @External
+    public void setStabilityFund(Address _address) {
+        onlyOwner();
+        isContract(_address);
+        stabilityFund.set(_address);
+    }
+
+    @External(readonly = true)
+    public Address getStabilityFund() {
+        return stabilityFund.get();
     }
 
     @External
@@ -203,6 +262,17 @@ public class FeeHandlerImpl implements FeeHandler {
                 transferToken(sender, getContractAddress("dividends"), getTokenBalance(sender), new byte[0]);
             }
         }
+        if (_from == loans.get()) {
+            loanFeesAccrued = loanFeesAccrued.add(_value);
+        }
+
+        if (_from == dex.get() && swapFeesAccruedDB.get(sender) != null) {
+            BigInteger accruedFees = swapFeesAccruedDB.get(sender);
+            swapFeesAccruedDB.set(sender, accruedFees.add(_value));
+        }
+        if (_from == stabilityFund.get()) {
+            stabilityFundFeesAccrued = stabilityFundFeesAccrued.add(_value);
+        }
     }
 
     @External
@@ -228,6 +298,17 @@ public class FeeHandlerImpl implements FeeHandler {
     public int getNextAllowedAddressIndex() {
         return nextAllowedAddressesIndex.getOrDefault(0);
     }
+
+    @External(readonly = true)
+    public BigInteger getLoanFeesAccrued() {
+        return loanFeesAccrued;
+    }
+
+    @External(readonly = true)
+    public BigInteger getStabilityFundFeesAccrued() {
+        return stabilityFundFeesAccrued;
+    }
+    // add swap fees getter
 
     @External
     public void route_contract_balances() {
